Use cobra to handle multiple commands.
* Rearranges to split commands from logic.
* Adds go.work
* Adds .gitignore to get rid of tuf stuff
