+++ .gitignore
@@ -1,0 +1,2 @@
+tuf-repo-cdn.sigstage.dev.json
+tuf-repo-cdn.sigstage.dev/
