+++ sourcetool/go.work
@@ -1,0 +1,3 @@
+go 1.24
+
+use .
